+++ src/main/java/org/example/studylog/controller/jwt/AuthController.java
@@ -4,7 +4,7 @@ import jakarta.servlet.http.Cookie;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import org.example.studylog.jwt.JWTUtil;
-import org.example.studylog.service.oauth.TokenService;
+import org.example.studylog.service.TokenService;
 import org.example.studylog.util.CookieUtil;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
