+++ src/main/java/org/example/studylog/service/TokenService.java
@@ -1,4 +1,4 @@
-package org.example.studylog.service.oauth;
+package org.example.studylog.service;
 
 import io.jsonwebtoken.ExpiredJwtException;
 import org.example.studylog.entity.RefreshEntity;
